Make nu_repl a testbin; Fix wrong cwd test error
