+++ tests/main.rs
@@ -2,7 +2,6 @@ extern crate nu_test_support;
 
 mod hooks;
 mod modules;
-mod nu_repl;
 mod overlays;
 mod parsing;
 mod path;
