+++ tests/overlays/mod.rs
@@ -1,5 +1,4 @@
-use super::nu_repl::nu_repl;
-use nu_test_support::{nu, pipeline};
+use nu_test_support::{nu, nu_repl_code, pipeline};
 
 #[test]
 fn add_overlay() {
@@ -10,7 +9,7 @@ fn add_overlay() {
     ];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert_eq!(actual.out, "foo");
     assert_eq!(actual_repl.out, "foo");
@@ -25,7 +24,7 @@ fn add_overlay_env() {
     ];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert_eq!(actual.out, "foo");
     assert_eq!(actual_repl.out, "foo");
@@ -36,7 +35,7 @@ fn add_overlay_from_file_decl() {
     let inp = &[r#"overlay add samples/spam.nu"#, r#"foo"#];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert_eq!(actual.out, "foo");
     assert_eq!(actual_repl.out, "foo");
@@ -47,7 +46,7 @@ fn add_overlay_from_file_decl() {
 fn add_overlay_from_file_decl_cd() {
     let inp = &[r#"cd samples"#, r#"overlay add spam.nu"#, r#"foo"#];
 
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert_eq!(actual_repl.out, "foo");
 }
@@ -57,7 +56,7 @@ fn add_overlay_from_file_alias() {
     let inp = &[r#"overlay add samples/spam.nu"#, r#"bar"#];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert_eq!(actual.out, "bar");
     assert_eq!(actual_repl.out, "bar");
@@ -68,7 +67,7 @@ fn add_overlay_from_file_env() {
     let inp = &[r#"overlay add samples/spam.nu"#, r#"$env.BAZ"#];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert_eq!(actual.out, "baz");
     assert_eq!(actual_repl.out, "baz");
@@ -83,7 +82,7 @@ fn add_overlay_scoped() {
     ];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert!(!actual.err.is_empty());
     #[cfg(windows)]
@@ -103,7 +102,7 @@ fn update_overlay_from_module() {
     ];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert_eq!(actual.out, "bar");
     assert_eq!(actual_repl.out, "bar");
@@ -120,7 +119,7 @@ fn update_overlay_from_module_env() {
     ];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert_eq!(actual.out, "bar");
     assert_eq!(actual_repl.out, "bar");
@@ -136,7 +135,7 @@ fn remove_overlay() {
     ];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert!(!actual.err.is_empty());
     #[cfg(windows)]
@@ -155,7 +154,7 @@ fn remove_last_overlay() {
     ];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert!(!actual.err.is_empty());
     #[cfg(windows)]
@@ -174,7 +173,7 @@ fn remove_overlay_scoped() {
     ];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert_eq!(actual.out, "foo");
     assert_eq!(actual_repl.out, "foo");
@@ -190,7 +189,7 @@ fn remove_overlay_env() {
     ];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert!(actual.err.contains("did you mean"));
     assert!(actual_repl.err.contains("DidYouMean"));
@@ -206,7 +205,7 @@ fn remove_overlay_scoped_env() {
     ];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert_eq!(actual.out, "foo");
     assert_eq!(actual_repl.out, "foo");
@@ -217,7 +216,7 @@ fn list_default_overlay() {
     let inp = &[r#"overlay list | last"#];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert_eq!(actual.out, "zero");
     assert_eq!(actual_repl.out, "zero");
@@ -232,7 +231,7 @@ fn list_last_overlay() {
     ];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert_eq!(actual.out, "spam");
     assert_eq!(actual_repl.out, "spam");
@@ -247,7 +246,7 @@ fn list_overlay_scoped() {
     ];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert_eq!(actual.out, "spam");
     assert_eq!(actual_repl.out, "spam");
@@ -263,7 +262,7 @@ fn remove_overlay_discard_decl() {
     ];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert!(!actual.err.is_empty());
     #[cfg(windows)]
@@ -282,7 +281,7 @@ fn remove_overlay_discard_alias() {
     ];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert!(!actual.err.is_empty());
     #[cfg(windows)]
@@ -295,13 +294,13 @@ fn remove_overlay_discard_alias() {
 fn remove_overlay_discard_env() {
     let inp = &[
         r#"overlay add samples/spam.nu"#,
-        r#"let-env BAGR = "bagr""#,
+        r#"let-env BAGR = `bagr`"#,
         r#"overlay remove spam"#,
         r#"$env.BAGR"#,
     ];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert!(actual.err.contains("did you mean"));
     assert!(actual_repl.err.contains("DidYouMean"));
@@ -317,7 +316,7 @@ fn remove_overlay_keep_decl() {
     ];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert!(actual.out.contains("bagr"));
     assert!(actual_repl.out.contains("bagr"));
@@ -327,13 +326,13 @@ fn remove_overlay_keep_decl() {
 fn remove_overlay_keep_alias() {
     let inp = &[
         r#"overlay add samples/spam.nu"#,
-        r#"alias bagr = "bagr""#,
+        r#"alias bagr = `bagr`"#,
         r#"overlay remove --keep-custom spam"#,
         r#"bagr"#,
     ];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert!(actual.out.contains("bagr"));
     assert!(actual_repl.out.contains("bagr"));
@@ -343,13 +342,13 @@ fn remove_overlay_keep_alias() {
 fn remove_overlay_keep_env() {
     let inp = &[
         r#"overlay add samples/spam.nu"#,
-        r#"let-env BAGR = "bagr""#,
+        r#"let-env BAGR = `bagr`"#,
         r#"overlay remove --keep-custom spam"#,
         r#"$env.BAGR"#,
     ];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert!(actual.out.contains("bagr"));
     assert!(actual_repl.out.contains("bagr"));
@@ -365,7 +364,7 @@ fn remove_overlay_keep_discard_overwritten_decl() {
     ];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert!(!actual.err.is_empty());
     #[cfg(windows)]
@@ -378,13 +377,13 @@ fn remove_overlay_keep_discard_overwritten_decl() {
 fn remove_overlay_keep_discard_overwritten_alias() {
     let inp = &[
         r#"overlay add samples/spam.nu"#,
-        r#"alias bar = 'baz'"#,
+        r#"alias bar = `baz`"#,
         r#"overlay remove --keep-custom spam"#,
         r#"bar"#,
     ];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert!(!actual.err.is_empty());
     #[cfg(windows)]
@@ -397,13 +396,13 @@ fn remove_overlay_keep_discard_overwritten_alias() {
 fn remove_overlay_keep_discard_overwritten_env() {
     let inp = &[
         r#"overlay add samples/spam.nu"#,
-        r#"let-env BAZ = "bagr""#,
+        r#"let-env BAZ = `bagr`"#,
         r#"overlay remove --keep-custom spam"#,
         r#"$env.BAZ"#,
     ];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert!(actual.err.contains("did you mean"));
     assert!(actual_repl.err.contains("DidYouMean"));
@@ -421,7 +420,7 @@ fn remove_overlay_keep_decl_in_latest_overlay() {
     ];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert!(actual.out.contains("bagr"));
     assert!(actual_repl.out.contains("bagr"));
@@ -431,7 +430,7 @@ fn remove_overlay_keep_decl_in_latest_overlay() {
 fn remove_overlay_keep_alias_in_latest_overlay() {
     let inp = &[
         r#"overlay add samples/spam.nu"#,
-        r#"alias bagr = 'bagr'"#,
+        r#"alias bagr = `bagr`"#,
         r#"module eggs { }"#,
         r#"overlay add eggs"#,
         r#"overlay remove --keep-custom spam"#,
@@ -439,7 +438,7 @@ fn remove_overlay_keep_alias_in_latest_overlay() {
     ];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert!(actual.out.contains("bagr"));
     assert!(actual_repl.out.contains("bagr"));
@@ -449,7 +448,7 @@ fn remove_overlay_keep_alias_in_latest_overlay() {
 fn remove_overlay_keep_env_in_latest_overlay() {
     let inp = &[
         r#"overlay add samples/spam.nu"#,
-        r#"let-env BAGR = "bagr""#,
+        r#"let-env BAGR = `bagr`"#,
         r#"module eggs { }"#,
         r#"overlay add eggs"#,
         r#"overlay remove --keep-custom spam"#,
@@ -457,7 +456,7 @@ fn remove_overlay_keep_env_in_latest_overlay() {
     ];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert!(actual.out.contains("bagr"));
     assert!(actual_repl.out.contains("bagr"));
@@ -474,7 +473,7 @@ fn preserve_overrides() {
     ];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert_eq!(actual.out, "new-foo");
     assert_eq!(actual_repl.out, "new-foo");
@@ -491,7 +490,7 @@ fn reset_overrides() {
     ];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert_eq!(actual.out, "foo");
     assert_eq!(actual_repl.out, "foo");
@@ -502,7 +501,7 @@ fn overlay_new() {
     let inp = &[r#"overlay new spam"#, r#"overlay list | last"#];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert_eq!(actual.out, "spam");
     assert_eq!(actual_repl.out, "spam");
@@ -518,7 +517,7 @@ fn overlay_keep_pwd() {
     ];
 
     let actual = nu!(cwd: "tests/overlays", pipeline(&inp.join("; ")));
-    let actual_repl = nu_repl("tests/overlays", inp);
+    let actual_repl = nu!(cwd: "tests/overlays", nu_repl_code(inp));
 
     assert_eq!(actual.out, "samples");
     assert_eq!(actual_repl.out, "samples");
